Remove build config script which is now replaced by the lambda function
