+++ server/cloudfront-gen-lambda.js
@@ -10,8 +10,6 @@
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE. See the License for more details.
  *
- * You should have received a copy of the License along with this program; if
- * not, you can obtain one from Tidepool Project at tidepool.org.
  */
 /* eslint-disable lodash/prefer-lodash-typecheck */
 const fs = require('fs');
@@ -20,12 +18,20 @@ const crypto = require('crypto');
 const handlebars = require('handlebars');
 const blipConfig = require('./config.app');
 
+const { ShellString } = require('shelljs');
+const _ = require('lodash');
+
+const reTitle = /<title>([^<]*)<\/title>/;
+const reZendesk = /(^\s+<!-- Start of support Zendesk Widget script -->\n)(.*\n)?(^\s+<!-- End of support Zendesk Widget script -->)/m;
+const reTrackerUrl = /const u = '(.*)';/;
+const reTrackerSiteId = /const id = ([0-9]);/;
+const reMatomoJs = /(^\s+<!-- Start of Tracker Code -->\n)(.*\n)*(^\s+<!-- End of Tracker Code -->)/m;
+const reCrowdin = /(^\s+<!-- Start of Crowdin -->\n)(.*\n)*(^\s+<!-- End of Crowdin -->)/m;
+const reCrowdinBranding = /BRANDING/;
+
 const reUrl = /(^https?:\/\/[^/]+).*/;
 const reDashCase = /[A-Z](?:(?=[^A-Z])|[A-Z]*(?=[A-Z][^A-Z]|$))/g;
-const scriptConfigJs = '<script defer type="text/javascript" src="config.js" integrity="{{CONFIG_HASH}}" crossorigin="anonymous"></script>';
-const templateFilename = 'templates/cloudfront-lambda-blip-request-viewer.js';
-const indexHtmlFilename = 'dist/index.html';
-const outputFilenameTemplate = 'dist/cloudfront-{{ TARGET_ENVIRONNEMENT }}-blip-request-viewer.js';
+const outputFilenameTemplate = '{{ DIST }}/cloudfront-{{ TARGET_ENVIRONNEMENT }}-blip-request-viewer.js';
 
 const featurePolicy = [
   "accelerometer 'none'",
@@ -63,6 +69,21 @@ let lambdaTemplate = null;
 let indexHtml = null;
 /** @type {string} */
 let distribFiles = null;
+
+let distDir = null;
+let srvDir = null;
+
+function getHash(str) {
+	const hash = crypto.createHash('md5');
+	hash.update(str);
+	return hash.digest('hex').substr(0, 20);
+}
+
+function getIntegrity(str, algorithm = 'sha512') {
+	const hash = crypto.createHash(algorithm);
+	hash.update(str);
+	return hash.digest('base64');
+}
 
 /**
  * Transform a camelCase string to dash-case
@@ -143,23 +164,19 @@ function genOutputFile() {
     return;
   }
 
-  const hash = crypto.createHash('sha512');
   let configJs = `window.config = ${JSON.stringify(blipConfig, null, 2)};`;
   console.log('Using config:', configJs);
-  hash.update(configJs);
-  const configHash = `sha512-${hash.digest('base64')}`;
-  console.log('Configuration hash:', configHash);
 
   const templateParameters = {
     ...blipConfig,
     DISTRIB_FILES: distribFiles,
     INDEX_HTML: '',
     CONFIG_JS: configJs,
-    CONFIG_HASH: configHash,
     TARGET_ENVIRONNEMENT: process.env.TARGET_ENVIRONNEMENT.toLowerCase(),
     FEATURE_POLICY: featurePolicy.join(';'),
     GEN_DATE: new Date().toISOString(),
     CSP: '',
+    DIST: distDir,
   };
 
   const csp = genContentSecurityPolicy();
@@ -213,9 +230,7 @@ function withIndexHtml(err, data) {
     process.exitCode = 1;
     return;
   }
-  console.log('Using', indexHtmlFilename);
 
-  indexHtml = data.replace(/<script[^>]+src="config(\.[0-9a-z]+)?.js"[^>]*><\/script>/, scriptConfigJs);
   indexHtml = indexHtml.replace(/<(script)/g, '<$1 nonce="${nonce}"');
   genOutputFile();
 }
@@ -235,14 +250,172 @@ function withTemplate(err, data) {
   genOutputFile();
 }
 
+
+/*** Main ***/
 if (typeof process.env.TARGET_ENVIRONNEMENT !== 'string' || process.env.TARGET_ENVIRONNEMENT.length < 1) {
   console.error('Missing environnement variable TARGET_ENVIRONNEMENT');
-  process.exitCode = 1;
+  process.exit(1);
 } else if (typeof process.env.API_HOST !== 'string' || !reUrl.test(process.env.API_HOST)) {
   console.error('Missing or invalid environnement variable API_HOST');
-  process.exitCode = 1;
-} else {
-  fs.readdir('dist', withFilesList);
-  fs.readFile(templateFilename, { encoding: 'utf-8' }, withTemplate);
-  fs.readFile(indexHtmlFilename, { encoding: 'utf-8' }, withIndexHtml);
+  process.exit(1);
 }
+
+// Determined dist dir location
+for (const dd of ['dist', '../dist']) {
+  distDir = dd;
+  if (fs.existsSync(distDir)) {
+    break;
+  }
+  distDir = null;
+}
+if (distDir === null) {
+  console.error('dist not found in . or ..');
+  process.exit(1);
+} else {
+  console.info(`Using dist directory: ${distDir}`);
+}
+
+// Determined server dir location:
+for (const sd of ['.', 'server']) {
+  srvDir = sd;
+  if (fs.existsSync(`${srvDir}/config.app.js`)) {
+    break;
+  }
+  srvDir = null;
+}
+if (srvDir === null) {
+  console.error('config.app.js not found, can\'t determined the server dir location.');
+  process.exit(1);
+} else {
+  console.info(`Using server directory: ${srvDir}`);
+}
+
+// Display configuration used
+console.info('Using configuration:', blipConfig);
+
+indexHtml = fs.readFileSync(`${distDir}/index.html`, 'utf8');
+if (typeof process.env.BRANDING === 'string') {
+  const title = process.env.BRANDING.replace(/^\w/, (c) => { return c.toUpperCase(); });
+  console.info(`- Setup title to ${title}`);
+  indexHtml = indexHtml.replace(reTitle, `<title>${title}</title>`);
+}
+
+// *** ZenDesk ***
+let helpLink = '<!-- Zendesk disabled -->';
+if (!reZendesk.test(indexHtml)) {
+  console.error(`/!\\ Can't find help pattern in index.html: ${reZendesk.source} /!\\`);
+  process.exit(1);
+}
+if (typeof process.env.HELP_LINK === 'string' && process.env.HELP_LINK.startsWith('https://')) {
+  console.info('- Using HELP_LINK:', process.env.HELP_LINK);
+  helpLink = `<script id="ze-snippet" type="text/javascript" defer src="${process.env.HELP_LINK}"></script>`;
+} else {
+  console.info('- Help link is disabled');
+}
+indexHtml = indexHtml.replace(reZendesk, `$1  ${helpLink}\n$3`);
+
+// *** Matomo ***
+if (!reMatomoJs.test(indexHtml)) {
+  console.error(`/!\\ Can't find tracker pattern in index.html: ${reMatomoJs.source} /!\\`);
+  process.exit(1);
+}
+switch (_.get(process, 'env.METRICS_SERVICE', 'disabled')) {
+case 'matomo':
+  console.info('- Using matomo tracker code');
+  if (!_.isEmpty(process.env.MATOMO_TRACKER_URL) && process.env.MATOMO_TRACKER_URL.startsWith('http')) {
+    // Replace tracker Javascript
+
+    let matomoTrackerUrl = process.env.MATOMO_TRACKER_URL;
+    if (!matomoTrackerUrl.endsWith('/')) {
+      matomoTrackerUrl = `${matomoTrackerUrl}/`;
+    }
+
+    let matomoJs = fs.readFileSync(`${srvDir}/templates/matomo.js`, 'utf8');
+    console.info(`  => Setting up matomo tracker code: ${matomoTrackerUrl}`);
+    const updatedSrc = matomoJs.replace(reTrackerUrl, (m, u) => {
+      return m.replace(u, matomoTrackerUrl);
+    });
+    const siteId = _.get(process, 'env.MATOMO_TRACKER_SITEID', 1);
+    matomoJs = updatedSrc.replace(reTrackerSiteId, (m, u) => {
+      return m.replace(u, siteId);
+    });
+
+    let fileHash = getHash(matomoJs);
+    const integrity = `integrity="sha512-${getIntegrity(matomoJs)}" crossorigin="anonymous"`;
+    const fileName = `matomo.${fileHash}.js`;
+
+    let matomoConfigScript = null;
+    let matomoScript = null;
+
+    // Public path declared (for CloudFront)
+    if (typeof process.env.PUBLIC_PATH === 'string' && process.env.PUBLIC_PATH.startsWith('https')) {
+      console.info(`  => Using public path: ${process.env.PUBLIC_PATH}`);
+      if (process.env.PUBLIC_PATH.endsWith('/')) {
+        matomoConfigScript = `<script defer type="text/javascript" src="${process.env.PUBLIC_PATH}${fileName}" ${integrity}></script>`;
+      } else {
+        matomoConfigScript = `<script defer type="text/javascript" src="${process.env.PUBLIC_PATH}/${fileName}" ${integrity}></script>`;
+      }
+    } else {
+      matomoConfigScript = `<script defer type="text/javascript" src="${fileName}" ${integrity}></script>`;
+    }
+
+    // Matomo main script
+    matomoScript = `<script defer type="text/javascript" src="${matomoTrackerUrl}matomo.js"></script>`;
+
+    const matomoConfigScripts = `  ${matomoConfigScript}\n  ${matomoScript}\n`;
+    indexHtml = indexHtml.replace(reMatomoJs, `$1${matomoConfigScripts}$3`);
+
+    const shellStr = new ShellString(matomoJs);
+    shellStr.to(`${distDir}/${fileName}`);
+  } else {
+    console.error('  /!\\ Invalid matomo config url, please verify your MATOMO_TRACKER_URL env variable /!\\');
+  }
+  break;
+case 'disabled':
+  console.info('- Tracker code is disabled');
+  indexHtml = indexHtml.replace(reMatomoJs, '$1  <!-- Tracker disabled -->\n$3');
+  break;
+default:
+  console.error(`/!\\ Unknown tracker ${process.env.METRICS_SERVICE} /!\\`);
+  indexHtml = indexHtml.replace(reMatomoJs, '$1  <!-- Tracker disabled -->\n$3');
+  break;
+}
+
+// *** Crowdin ***
+if (process.env.CROWDIN === 'enabled') {
+  console.info('- Enable crowdin...');
+  let crowdinJs = fs.readFileSync(`${srvDir}/templates/crowdin.js`, 'utf8');
+  let crowdinProject = blipConfig.BRANDING;
+  switch (blipConfig.BRANDING) {
+  case 'diabeloop':
+    crowdinProject = 'yourloops';
+    break;
+  }
+  console.info(`  => Setting up crowdin project: ${crowdinProject}`);
+  crowdinJs = crowdinJs.replace(reCrowdinBranding, crowdinProject);
+  const fileHash = getHash(crowdinJs);
+  const integrity = getIntegrity(crowdinJs);
+  const fileName = `crowdin.${fileHash}.js`;
+  const shellStr = new ShellString(crowdinJs);
+  shellStr.to(`${distDir}/${fileName}`);
+
+  const crowdinScripts = `\
+  <script type="text/javascript" defer src="${fileName}" integrity="sha512-${integrity}" crossorigin="anonymous"></script>\n\
+  <script type="text/javascript" defer src="https://cdn.crowdin.com/jipt/jipt.js"></script>`;
+  if (!reCrowdin.test(indexHtml)) {
+    console.error(`/!\\ Can't find crowdin pattern in index.html: ${reCrowdin.source} /!\\`);
+    process.exit(1);
+  }
+  indexHtml = indexHtml.replace(reCrowdin, `$1${crowdinScripts}\n$3`);
+} else {
+  console.info('- Crowdin is disabled');
+  indexHtml = indexHtml.replace(reCrowdin, '$1  <!-- disabled -->\n$3');
+}
+
+const templateFilename = 'templates/cloudfront-lambda-blip-request-viewer.js';
+
+
+fs.readdir(distDir, withFilesList);
+fs.readFile(templateFilename, { encoding: 'utf-8' }, withTemplate);
+fs.readFile(`${distDir}/index.html`, { encoding: 'utf-8' }, withIndexHtml);
+
